% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_gshhg.R
\docType{data}
\name{Bounding box data}
\alias{Bounding box data}
\alias{alaska_bbox}
\alias{bering_bbox}
\alias{calcur_bbox}
\alias{hawaii_bbox}
\alias{npac_bbox}
\alias{us_arctic_bbox}
\title{This is data to be included in my package}
\description{
Spatial bounding boxes used for extracting specific geographic reagional data.
}
\keyword{data}
