% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_gshhg.R
\name{extract_gshhg}
\alias{extract_gshhg}
\alias{bering}
\alias{alaska}
\alias{us_arctic}
\alias{npac}
\alias{calcur}
\alias{hawaii}
\title{Extract Region from GSHHG}
\usage{
extract_gshhg(
  data,
  resolution = "i",
  epsg = NULL,
  buffer = 5000,
  simplify = FALSE,
  warn = FALSE
)

bering(resolution = "i", epsg = 3571, simplify = FALSE)

alaska(resolution = "i", epsg = 3338, simplify = FALSE)

us_arctic(resolution = "i", epsg = 3572, simplify = FALSE)

npac(resolution = "i", epsg = 3832, simplify = FALSE)

calcur(resolution = "i", epsg = 3310, simplify = FALSE)

hawaii(resolution = "i", epsg = 3857, simplify = FALSE)
}
\arguments{
\item{data}{an sf object that defines that area of interest}

\item{resolution}{either "f", "h", "i", or "c"}

\item{epsg}{integer indicating the numeric epsg value (e.g. 3571)}

\item{buffer}{integer indicating a value in projected units to buffer data}

\item{simplify}{TRUE/FALSE whether to call rmapshaper::ms_simplify}

\item{warn}{TRUE/FALSE turn off projection warnings given by spatial functions. These warnings can generally be ignored.}
}
\description{
Functions to extract specified regions from the GSHHG dataset

This map region is centered on the Bering Sea. Coordinates are returned in 
the Lambert-Azimuthal Equal Area Bering Sea Projection (epsg:3571)

This map region covers the extent of Alaska. Coordinates are returned in the
Alaska Albers Projection (epsg:3338)

This map region covers the US Arctic region centered on northern Alaska. The
region extends into the Russian East Siberian Sea and the Eastern Beaufort 
Sea of Canada. Coordinates are returned in the Lambert-Azimuthal Equal Area
Alaska Projection (epsg:3572).

This map region covers the entire North Pacific Ocean from Hawaii in the
south to the Bering Sea in the north. Coordinates are returned in the PDC
Mercator Projection (epsg:3832)

This map region covers the California current off of the west coast of the 
United States. The region extends from southern British Columbia, Canada to
the Baja Peninsula of Mexico. Coordinates are returned in the California
(Teale) Albers Projection (epsg:3310).

This map region covers the extent of Hawaii. Coordinates are returned in the
WGS 84 / Pseudo-Mercator -- Spherical Mercator (epsg:3857)
}
