% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortifi.R
\name{fortifi}
\alias{fortifi}
\title{fortifi}
\usage{
fortifi(poly, tol, minarea = NA)
}
\arguments{
\item{poly}{SpatialPolygon.}

\item{tol}{tolerance passed to the rgeos::gSimplify() function}

\item{minarea}{minimum area cutoff for printing a polygon}
}
\value{
fortified poly as a data.frame ready for ggplot
}
\description{
a customized fortify function to optionally reduce detail when
plotting large shapefiles in ggplot
}
\author{
Josh M London
}
