% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAKCRM.bathy.R
\name{getAKCRM.bathy}
\alias{getAKCRM.bathy}
\title{getAKCRM.bathy}
\usage{
getAKCRM.bathy(lon1, lon2, lat1, lat2, resolution = 0.4, keep = FALSE)
}
\arguments{
\item{lon1}{}

\item{keep}{}
}
\value{
bathy class object
}
\description{
a modification of the getNOAA.bathy() function in the marmap package
}
\details{
this allows one to pull bathymetry data from the South Alaska CRM
24-arc second resolution dataset for Alaska. Unlike getNOAA.bathy(),
the bounding box coordinates should be from 0-360
}
\author{
Josh M London
}

