% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggExpansion.R
\name{ggExpansion}
\alias{ggExpansion}
\title{ggExpansion}
\usage{
ggExpansion(df, x = x, y = y, x_fac, y_fac, min_dist = 10000)
}
\description{
a functions for expanding the xlim and ylim ggplot values
}
\details{
this allows the user to provide an expansion factor for xlim
and ylim values in ggplot
}

