% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_gshhg.R
\name{install_gshhg}
\alias{install_gshhg}
\title{Download and Install GSHHG Binary Data}
\usage{
install_gshhg()
}
\description{
This is a function for downloading and installing the latest version of the
GSHHG binary dataset. Current version supported is 2.3.7. This only needs
to be run once after the package is installed.
}
