% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/nPacMaps.R
\docType{package}
\name{nPacMaps}
\alias{nPacMaps}
\alias{nPacMaps-package}
\title{Map data (and a few functions) for the North Pacific}
\description{
Map data (and a few functions) for the North Pacific
}

