% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nPacMaps.R
\docType{package}
\name{ptolemy}
\alias{ptolemy}
\title{ptolemy: an R package for accessing global high-resolution geography}
\description{
Provides easy access to the GSHHG data base of coastline data from NOAA.
}
