% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/to_km.R
\name{to_km}
\alias{to_km}
\title{to_km}
\usage{
to_km(pos.only = FALSE)
}
\description{
convert tick labels in ggplot from meters to km. designed for use within
the scale functions of ggplot2
}
\author{
Josh M London
}

