% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_gshhg.R
\name{extract_gshhg}
\alias{extract_gshhg}
\alias{bering}
\alias{alaska}
\alias{us_arctic}
\alias{npac}
\alias{calcur}
\title{Extract Region from GSHHG}
\usage{
extract_gshhg(xlims, ylims, resolution = "i", epsg, fortify = TRUE)

bering(xlims = c(180 - 50, 180 + 55), ylims = c(35, 80), resolution = "i",
  epsg = "3571", fortify = TRUE)

alaska(xlims = c(180 - 5, 180 + 50), ylims = c(35, 75), resolution = "i",
  epsg = "3338", fortify = TRUE)

us_arctic(xlims = c(180 - 2, 180 + 50), ylims = c(60, 90),
  resolution = "i", epsg = "3572", fortify = TRUE)

npac(xlims = c(180 - 50, 180 + 70), ylims = c(20, 67), resolution = "i",
  epsg = "3832", fortify = TRUE)

calcur(xlims = c(180 + 35, 180 + 70), ylims = c(26, 52), resolution = "i",
  epsg = "3310", fortify = TRUE)
}
\arguments{
\item{xlims}{a vector of x coordinate limits; 0-360 degrees}

\item{ylims}{a vector of y coordinate limits: 0-90 degrees}

\item{resolution}{either "f", "h", "i", or "c"}

\item{epsg}{character indicating the numeric epsg value (e.g. "3571")}

\item{fortify}{TRUE/FALSE whether to return a fortified data.frame for ggplot (now relies on broom::tidy instead of ggplot2::fortify())}
}
\description{
Functions to extract specified N. Pacific regions from the GSHHG dataset

This map region is centered on the Bering Sea. Coordinates are returned in 
the Lambert-Azimuthal Equal Area Bering Sea Projection (epsg:3571)

This map region covers the extent of Alaska. Coordinates are returned in the
Alaska Albers Projection (epsg:3338)

This map region covers the US Arctic region centered on northern Alaska. The
region extends into the Russian East Siberian Sea and the Eastern Beaufort 
Sea of Canada. Coordinates are returned in the Lambert-Azimuthal Equal Area
Alaska Projection (epsg:3572).

This map region covers the entire North Pacific Ocean from Hawaii in the
south to the Bering Sea in the north. Coordinates are returned in the PDC
Mercator Projection (epsg:3832)

This map region covers the California current off of the west coast of the 
United States. The region extends from southern British Columbia, Canada to
the Baja Peninsula of Mexico. Coordinates are returned in the California
(Teale) Albers Projection (epsg:3310).
}
