% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nPacMaps.R
\docType{package}
\name{ptolemy}
\alias{ptolemy}
\alias{ptolemy-package}
\title{ptolemy: an R package for accessing global high-resolution geography}
\description{
ptolemy: an R package for accessing global high-resolution geography
}
